requirements.txt
BeautifulSoup4
html5lib
requests
irc
chardet
Pillow
